+++ Engine/Source/Core/Input.cpp
@@ -1,5 +1,10 @@
 #include "Input.h"
+#include "Logger.h"
+#include "Engine.h"
+#include "Math/Math.h"
+
 #include <unordered_map>
+#include <GLFW/glfw3.h>
 
 namespace Core
 {
@@ -8,10 +13,19 @@ namespace Core
         bool init;
         std::unordered_map<Keys, bool> keys;
         std::unordered_map<Buttons, bool> buttons;
+        std::unordered_map<Keys, bool> justKeys;
+        std::unordered_map<Keys, bool> lastKeys;
 
         Vector2 mousePos;
         Vector2 deltaPos;
         Vector2 lastPos;
+
+        MouseMode lastMode;
+        MouseMode mode;
+
+        float lastScroll;
+        float deltaScroll;
+        bool negativeDelta = false;
     };
 
     static InputState state;
@@ -47,15 +61,74 @@ namespace Core
 
         return state.buttons[button] != false;
     }
-    
+
+    void Input::SetMouseMode(MouseMode mode)
+    {
+        if (state.mode == mode)
+            return;
+        state.mode = mode;
+        if (state.lastMode != state.mode)
+        {
+            state.lastMode = state.mode;
+            Window *window = Engine::GetWindow();
+            if (!window)
+                return;
+
+            switch (state.mode)
+            {
+            case MouseMode::Visible:
+                glfwSetInputMode(window->GetHandle(), GLFW_CURSOR, GLFW_CURSOR_NORMAL);
+                break;
+
+            case MouseMode::Hidden:
+                glfwSetInputMode(window->GetHandle(), GLFW_CURSOR, GLFW_CURSOR_HIDDEN);
+                break;
+
+            case MouseMode::Locked:
+                glfwSetInputMode(window->GetHandle(), GLFW_CURSOR, GLFW_CURSOR_DISABLED);
+                break;
+
+            default:
+                break;
+            }
+        }
+    }
+
     Vector2 Input::GetMousePosition()
     {
         return state.mousePos;
     }
-    
+
     Vector2 Input::GetMouseDelta()
     {
         return state.deltaPos;
+    }
+
+    void Input::SetMousePosition(const Vector2 &vector)
+    {
+        Window *win = Engine::GetWindow();
+        if (!win)
+            return;
+
+        glfwSetCursorPos(win->GetHandle(), vector.x, vector.y);
+        state.mousePos = vector;
+    }
+
+    bool Input::GetKeyJustNow(Keys key)
+    {
+        if (!state.init)
+            return false;
+
+        return state.justKeys[key];
+    }
+
+    float Input::GetMouseWheelDelta()
+    {
+        return state.deltaScroll;
+    }
+
+    void Input::InternalUpdate()
+    {
     }
 
     void InputUpdateKey(Keys key, bool pressed)
@@ -63,7 +136,9 @@ namespace Core
         if (!state.init)
             return;
 
+        state.lastKeys[key] = state.keys[key];
         state.keys[key] = pressed;
+        state.justKeys[key] = (state.keys[key] && !state.lastKeys[key]);
     }
 
     void InputUpdateButton(Buttons button, bool pressed)
@@ -82,5 +157,30 @@ namespace Core
         state.deltaPos.Set(x - state.lastPos.x, y - state.lastPos.y);
         state.mousePos.Set(x, y);
         state.lastPos.Set(x, y);
+    }
+
+    void InputUpdateScroll(int x, int y)
+    {
+        if (y == 0)
+        {
+            state.deltaScroll = 0;
+            state.lastScroll = 0;
+            return;
+        }
+
+        state.deltaScroll = 0;
+
+        if (y < 0)
+        {
+            state.deltaScroll = y - Math::Positive(state.lastScroll);
+            state.lastScroll = Math::Positive(y);
+            state.negativeDelta = -1;
+        }
+        else
+        {
+            state.deltaScroll = y + Math::Positive(state.lastScroll);
+            state.lastScroll = Math::Positive(y);
+            state.negativeDelta = 1;
+        }
     }
 }