+++ Engine/Source/Math/Math.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "Base.h"
+
+namespace Core
+{
+    namespace Math
+    {
+        CE_API float Positive(float x);
+        CE_API float Clamp(float value, float min, float max);
+    };
+}