+++ Engine/Source/Renderer/Camera/Camera.cpp
@@ -8,7 +8,11 @@ namespace Core
 
     void Camera::CalculateView()
     {
-        view = Matrix4::Translate(0, 0, 0);
+
+        if (transformType == TransformationType::Transform)
+            matrixTransform = transform.GetTransformMatrix();
+
+        view.From(&matrixTransform);
     }
 
     void Camera::SetOriginPoint(OriginPoint o)
@@ -23,5 +27,32 @@ namespace Core
     void Camera::SetType(Type o)
     {
         type = o;
+    }
+
+    void Camera::SetTransformationType(TransformationType o)
+    {
+        if (transformType == o)
+            return;
+
+        if (transformType == TransformationType::Transform)
+        {
+            if (o == TransformationType::Matrix)
+            {
+                transform.Position = transform.Rotation = {0, 0, 0};
+                transform.Scale = {1, 1, 1};
+            }
+        }
+
+        transformType = o;
+    }
+
+    void Camera::SetZoom(float z)
+    {
+        zoom = z;
+    }
+
+    void Camera::AddZoom(float z)
+    {
+        zoom += z;
     }
 }